refactor wave51 runner and sandbox splits
