+++ crates/assay-core/src/engine/runner_next/mod.rs
@@ -3,10 +3,12 @@
 //! `src/engine/runner.rs` remains the stable facade; implementation blocks are
 //! mechanically moved here.
 
+pub(crate) mod assertions;
 pub(crate) mod baseline;
 pub(crate) mod cache;
 pub(crate) mod errors;
 pub(crate) mod execute;
 pub(crate) mod retry;
 pub(crate) mod scoring;
+pub(crate) mod single;
 pub(crate) mod tests;
